+++ backend/src/modules/transactional_outbox/index.js
@@ -1,0 +1,1 @@
+module.exports={...require('./OutboxRepository'),...require('./OutboxWorker'),...require('./OutboxAdminService')};